finished testing and bugfixing, this draft ready for merge

+++ validator_cleanup.rb
@@ -21,7 +21,7 @@ tmp_dir=File.join(working_dir, basename_normalized)
 working_file = File.join(tmp_dir, filename_normalized)
 bookinfo_file = File.join(tmp_dir,'book_info.json')
 stylecheck_file = File.join(tmp_dir,'style_check.json')
-submitter_file = File.join(tmp_dir,'submitter.json')
+submitter_file = File.join(tmp_dir,'contact_info.json')
 testing_value_file = File.join("C:", "staging.txt")
 inprogress_file = File.join(inbox,"#{filename_normalized}_IN_PROGRESS.txt")
 errFile = File.join(inbox, "ERROR_RUNNING_#{filename_normalized}.txt")
@@ -82,21 +82,21 @@ when File.file?(errFile)
 	logger.info('validator_cleanup') {"errFile found, indicating file failed basic validation, moving orig & errNotice from IN to OUT folder"}
 	FileUtils.mv input_file, outbox
 	FileUtils.mv errFile, outbox
-when !File.file?(bookinfo_file) || !stylecheck_styled
+when !File.file?(bookinfo_file) || (File.file?(stylecheck_file) && !stylecheck_styled)
 	FileUtils.mv input_file, outbox	  #return the original file to user
-	FileUtils.rm_rf tmp_dir   #alt:  could keep the tmpdir for review like following case
-	if !stylecheck_styled
-		logger.info('validator_cleanup') {"adding warn notice to outbox for unstyled doc"}
-		File.open(warn_notice, 'w') { |f|
-        	f.puts "WARNING: Your document is NOT STYLED. Bookmaker_validator cannot run on an unstyled document."
-    	}
-    else
+	if !File.file?(bookinfo_file)
     	logger.info('validator_cleanup') {"book_info.json missing, returned orig file and isbn lookup error notice to user, exiting cleanup"}
     	File.open(err_notice, 'w') { |f|
     		f.puts "isbn lookup failed for file #{filename_normalized}, bookmaker_validator could not run! Please double-check the isbn in your filename!"
     	}
-	end 
-when errlog || !stylecheck_complete
+    else
+		logger.info('validator_cleanup') {"adding warn notice to outbox for unstyled doc"}
+		File.open(warn_notice, 'w') { |f|
+        	f.puts "WARNING: Your document is NOT STYLED. Bookmaker_validator cannot run on an unstyled document."
+    	}
+	end
+	FileUtils.rm_rf tmp_dir   #alt:  could keep the tmpdir for review like following case	
+when errlog || (File.file?(stylecheck_file) && !stylecheck_complete)
 	logger.info('validator_cleanup') {"a major error(ALERT) was detected while running macros on \"#{filename_normalized}\", moving tmpdir to logfolder for further study"}
 	FileUtils.mv tmp_dir, "#{tmp_dir}__#{timestamp}"  #rename folder
 	FileUtils.mv "#{tmp_dir}__#{timestamp}", logfolder 
@@ -111,13 +111,13 @@ else
 	FileUtils.mv done_file, outbox
 	FileUtils.mv input_file, outbox
 	FileUtils.rm_rf tmp_dir
-	if isbn_mismatch
-	 	logger.info('validator_cleanup') {"adding warn notice to outbox for isbn mismatch"}
-	 	File.open(warn_notice, 'w') { |f|
-         	f.puts "WARNING: ISBN mismatch: the ISBN in your document's filename does not match one found in the manuscript."
-     	}
-	end 
 	logger.info('validator_cleanup') {"processing of \"#{filename_normalized}\" completed"}
 end
+if isbn_mismatch
+ 	logger.info('validator_cleanup') {"adding warn notice to outbox for isbn mismatch"}
+ 	File.open(warn_notice, 'a') { |f|
+       	f.puts "WARNING: ISBN mismatch: the ISBN in your document's filename does not match one found in the manuscript."
+   	}
+end 
 
 if File.file?(inprogress_file) then FileUtils.rm inprogress_file end
